Revert "float get_groundlevelpop"
This reverts commit fd997523c1fe012315e5e38b4995cd964a32f794.

+++ ltepop.h
@@ -3,7 +3,7 @@
 
 #include <vector>
 
-[[nodiscard]] auto get_groundlevelpop(int nonemptymgi, int element, int ion) -> float;
+[[nodiscard]] auto get_groundlevelpop(int nonemptymgi, int element, int ion) -> double;
 
 [[nodiscard]] auto calculate_levelpop(int nonemptymgi, int element, int ion, int level) -> double;
